Integrate Setting into Index (should be client)
